fixes and platform support
